requirements.txt
pandas
numpy
seaborn
scikit-learn
pdfplumber
python-docx
spacy
nltk

-e .